correct spots
