+++ src/lib/actions/mouse-position.ts
@@ -1,0 +1,34 @@
+import { inView } from 'motion';
+import { type Writable, writable } from 'svelte/store';
+
+export const useMousePosition = () => {
+    let position = writable<{ x: number; y: number }>({
+        x: 0,
+        y: 0
+    });
+
+    const action = (node: HTMLElement) => {
+        const handleMouseMove = (event: MouseEvent) => {
+            position.set({
+                x: event.clientX,
+                y: event.clientY
+            });
+        };
+
+        inView(
+            node,
+            () => {
+                node.addEventListener('mousemove', handleMouseMove);
+            },
+            { amount: 'all' }
+        );
+
+        return {
+            destroy() {
+                node.removeEventListener('mousemove', handleMouseMove);
+            }
+        };
+    };
+
+    return { action, position };
+};
